moved app/utils.go to protocols/cosi
